requirements/requirements.txt
Django>=3.0,<4.0
Jinja2==3.0.1
MarkupSafe==2.0.1
PyMySQL==1.0.2
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
djangorestframework==3.12.4
drf-spectacular==0.17.3
idna==3.2
itypes==1.2.0
mysqlclient==2.0.3
openapi-codec==1.3.2
parse-type==0.5.2
parse==1.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
uwsgi>2<2.1
six==1.16.0
mock
isort==5.9.3
mypy
pylint
flake8
black==21.7b0
coverage==5.5
safety
bandit==1.7.0
django-extensions==3.1.3